Fixed compiler warnings.
git-svn-id: http://chianti.ucsd.edu/svn@18899 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsd/ruschein/DenovoPGNetworkAlignment/src/data/FloatVector.java
@@ -250,7 +250,7 @@ public class FloatVector extends DataVector {
 	}
 
 	public float getAsFloat(int i) {
-		return (float) get(i);
+		return get(i);
 	}
 
 	public float get(int i) {
